+++ module/core/database/migration/migrations/mysql/000009_products_gallery_variation_position.down.sql
@@ -1,0 +1,4 @@
+DROP INDEX idx_product_gallery_items_variation_position ON product_gallery_items;
+
+ALTER TABLE product_gallery_items
+    DROP COLUMN variation_position;